+++ engine/render/source/render/particles/generators/alpha_particle_system_generator.cpp
@@ -9,8 +9,8 @@ namespace kw {
 
 ParticleSystemGenerator* AlphaParticleSystemGenerator::create_from_markdown(MemoryResource& memory_resource, ObjectNode& node) {
     return memory_resource.construct<AlphaParticleSystemGenerator>(
-        static_cast<float>(node["min"].as<NumberNode>().get_value()),
-        static_cast<float>(node["max"].as<NumberNode>().get_value())
+        node["min"].as<NumberNode>().get_value(),
+        node["max"].as<NumberNode>().get_value()
     );
 }
